requirements.txt
# AI API
anthropic>=0.34.0
openai>=1.0.0

# 環境変数管理
python-dotenv>=1.0.0

# Google APIs
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# HTTP通信
requests>=2.31.0

google-cloud-storage
